# Improve the performance of consecutive range-constrained PRF evaluations