--- a/bench/bench_rcprf.cpp
+++ b/bench/bench_rcprf.cpp
@@ -22,6 +22,8 @@
 
 #include <benchmark/benchmark.h>
 
+#include <random>
+
 using sse::crypto::Key;
 using sse::crypto::RCPrf;
 using sse::crypto::RCPrfParams;
@@ -31,17 +33,50 @@ static void BM_RCPrf_eval_all_loop(benchmark::State& state)
     uint8_t  depth          = state.range(0);
     uint64_t max_leaf_index = RCPrfParams::max_leaf_index_generic(depth);
 
+    std::random_device                      rnd;
+    std::mt19937_64                         rnd_gen(rnd());
+    std::uniform_int_distribution<uint64_t> unif_dist(
+        0, max_leaf_index - state.range(1));
+
     RCPrf<32> rcprf(Key<RCPrfParams::kKeySize>(), depth);
 
-    size_t i = 0;
     for (auto _ : state) {
-        rcprf.eval(i);
-
-        i = (i + 1) % (max_leaf_index + 1);
+        // randomly generate a starting point
+        uint64_t start_index = unif_dist(rnd_gen);
+        for (uint64_t i = 0; i < static_cast<uint64_t>(state.range(1)); i++) {
+            rcprf.eval(start_index + i);
+        }
     }
-    state.SetItemsProcessed(state.iterations());
+    state.SetItemsProcessed(state.iterations() * state.range(1));
+}
+
+static void BM_RCPrf_eval_all_range(benchmark::State& state)
+{
+    uint8_t  depth          = state.range(0);
+    uint64_t max_leaf_index = RCPrfParams::max_leaf_index_generic(depth);
+
+    std::random_device                      rnd;
+    std::mt19937_64                         rnd_gen(rnd());
+    std::uniform_int_distribution<uint64_t> unif_dist(
+        0, max_leaf_index - state.range(1));
+
+    RCPrf<32> rcprf(Key<RCPrfParams::kKeySize>(), depth);
+
+    auto callback = [](size_t, std::array<uint8_t, 32>) {};
+
+    for (auto _ : state) {
+        // randomly generate a starting point
+        uint64_t start_index = unif_dist(rnd_gen);
+
+        rcprf.eval_range(start_index, start_index + state.range(1), callback);
+    }
+    state.SetItemsProcessed(state.iterations() * state.range(1));
 }
 
 BENCHMARK(BM_RCPrf_eval_all_loop)
     ->RangeMultiplier(2)
-    ->Range(4, 62); // Arg(4)->Arg(8)->Arg(16)->Arg(4)->Arg(4);
+    ->Ranges({{16, 32}, {8, 128}});
+
+BENCHMARK(BM_RCPrf_eval_all_range)
+    ->RangeMultiplier(2)
+    ->Ranges({{16, 32}, {8, 128}});