… tests/test_rcprf.cpp