To evaluate a RCPrf (or ConstrainedRCPrf) object on consecutive inputs, we can only use a loop, and call RCPrf::eval in the body of this loop. This is quite inefficient as similar computations can be performed several times. Indeed, for a RCPrf based on a tree of depth `d`, it requires `N*d` PRG evaluations.

This PR implements the efficient evaluation of a RCPrf on a range, using a standard recursion technique which ensures that no Prg evaluation is done twice. As usual, this can be seen as a time-memory tradeoff (the used memory in that case is the stack). It can result in a `log N` speedup in the best case (however, this needs to be checked).

In practice, the improvement is big, as can be seen by running the `benchmark_rcprf` benchmark. My run follows:
```
----------------------------------------------------------------------------
Benchmark                                     Time           CPU Iterations
----------------------------------------------------------------------------
BM_RCPrf_eval/48                         693495 ns     692542 ns       1004   1.41011k items/s
BM_RCPrf_eval_range/48/8                 831000 ns     830829 ns        835   9.40326k items/s
BM_RCPrf_eval_range/48/16                994640 ns     993869 ns        733   15.7214k items/s
BM_RCPrf_eval_range/48/32               1216956 ns    1216353 ns        575   25.6916k items/s
BM_RCPrf_eval_range/48/64               1662742 ns    1662546 ns        421   37.5929k items/s
BM_RCPrf_eval_range/48/128              2688190 ns    2687935 ns        261   46.5041k items/s
```

As expected, the speedup increases with the range's size (the second number in the benchmarks' names). Similar results can be obtained after the PRF has been constrained (the figures do not account for the constrain time):
```
----------------------------------------------------------------------------
Benchmark                                     Time           CPU Iterations
----------------------------------------------------------------------------
BM_RCPrf_eval_range_constrain/48/8        78154 ns      78063 ns       8691   100.079k items/s
BM_RCPrf_eval_range_constrain/48/16      181070 ns     180989 ns       3857   86.3313k items/s
BM_RCPrf_eval_range_constrain/48/32      404702 ns     404196 ns       1753    77.314k items/s
BM_RCPrf_eval_range_constrain/48/64      888989 ns     888785 ns        809   70.3207k items/s
BM_RCPrf_eval_range_constrain/48/128    1820928 ns    1820510 ns        384   68.6621k items/s
```